refactor(cli): address review on hook-stdin + hook test fixtures
- hook-stdin: make wrappedRead never reject - a stdin stream failure (or the
  for-await iterator rejecting on timeout-driven destroy()) now degrades to the
  same quiet no-op as an empty stream, upholding readObservedHarnessSessionId's
  never-throws contract instead of bubbling out of the hook.
- extract the duplicated hook-command test scaffolding (makeCtx, stubStdin,
  agent-identity env save/restore) into a shared hook-test-helpers module.
